Spring Rest - Restrict input for POST /api/evetns

+++ SpringRest/src/main/java/me/ewan/springrestapi/evetns/EventRepository.java
@@ -1,6 +1,8 @@
 package me.ewan.springrestapi.evetns;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 }
